Simulated actions (WIP)
Agent makes simulated actions before committing to an action.  Movements, death are rolled back after a simulation step.  Pickups are not yet rolled back.  Agent color changes to display whether it is in a simulated or acting state.  ObjectCollider script removed & remaining objects prefabs restructured so that their collider and sprite are on the main object.